# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BillCreate": {
        "additionalProperties": false,
        "properties": {
          "amount_max": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Amount Max"
          },
          "amount_min": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Amount Min"
          },
          "currency_code": {
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency Code",
            "type": "string"
          },
          "first_due_date": {
            "format": "date",
            "title": "First Due Date",
            "type": "string"
          },
          "frequency": {
            "$ref": "#/components/schemas/BillFrequency"
          },
          "match_text": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Text"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "currency_code",
          "amount_min",
          "amount_max",
          "first_due_date",
          "frequency"
        ],
        "title": "BillCreate",
        "type": "object"
      },
...
      "BillOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "amount_max": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
            "title": "Amount Max",
            "type": "string"
          },
          "amount_min": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
            "title": "Amount Min",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency_code": {
            "title": "Currency Code",
            "type": "string"
          },
          "first_due_date": {
            "format": "date",
            "title": "First Due Date",
            "type": "string"
          },
          "frequency": {
            "$ref": "#/components/schemas/BillFrequency"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "match_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Text"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "currency_code",
          "amount_min",
          "amount_max",
          "match_text",
          "first_due_date",
          "frequency",
          "active",
          "created_at"
        ],
        "title": "BillOut",
        "type": "object"
      },
      "BillStatusOut": {
        "description": "A conta a pagar com a situacao na data pedida.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "amount_max": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
            "title": "Amount Max",
            "type": "string"
          },
          "amount_min": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
            "title": "Amount Min",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency_code": {
            "title": "Currency Code",
            "type": "string"
          },
          "first_due_date": {
            "format": "date",
            "title": "First Due Date",
            "type": "string"
          },
          "frequency": {
            "$ref": "#/components/schemas/BillFrequency"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Due Date"
          },
          "match_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Text"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "next_due_date": {
            "format": "date",
            "title": "Next Due Date",
            "type": "string"
          },
          "next_due_paid": {
            "title": "Next Due Paid",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "upcoming",
              "paid",
              "overdue"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "currency_code",
          "amount_min",
          "amount_max",
          "match_text",
          "first_due_date",
          "frequency",
          "active",
          "created_at",
          "last_due_date",
          "next_due_date",
          "status",
          "next_due_paid"
        ],
        "title": "BillStatusOut",
        "type": "object"
      },
      "BillUpdate": {
        "additionalProperties": false,
        "description": "A moeda nao muda. `match_text` enviado como null apaga a ligacao automatica.",
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "amount_max": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Max"
          },
          "amount_min": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Min"
          },
          "first_due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Due Date"
          },
          "frequency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BillFrequency"
              },
              {
                "type": "null"
              }
            ]
          },
          "match_text": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Text"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "BillUpdate",
        "type": "object"
      },
      "BudgetCreate": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "currency_code": {
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency Code",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/BudgetPeriod"
          }
        },
        "required": [
          "name",
          "currency_code",
          "amount",
          "period"
        ],
        "title": "BudgetCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_BillOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BillOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[BillOut]",
        "type": "object"
      },
      "Page_BudgetOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BudgetOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[BudgetOut]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "finance-app",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/bills": {
      "get": {
        "operationId": "list_bills_api_v1_bills_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BillOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Bills",
        "tags": [
          "bills"
        ]
      },
      "post": {
        "operationId": "create_bill_api_v1_bills_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Bill",
        "tags": [
          "bills"
        ]
      }
    },
    "/api/v1/bills/status": {
      "get": {
        "description": "Ultimo e proximo vencimento de cada conta a pagar na data `on` (hoje, se omitida), e se o\nultimo foi pago. Ordenadas pelo proximo vencimento.",
        "operationId": "bills_status_api_v1_bills_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "on",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "On"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BillStatusOut"
                  },
                  "title": "Response Bills Status Api V1 Bills Status Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Bills Status",
        "tags": [
          "bills"
        ]
      }
    },
    "/api/v1/bills/{bill_id}": {
      "delete": {
        "operationId": "delete_bill_api_v1_bills__bill_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "bill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Bill",
        "tags": [
          "bills"
        ]
      },
      "get": {
        "operationId": "get_bill_api_v1_bills__bill_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "bill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Bill",
        "tags": [
          "bills"
        ]
      },
      "patch": {
        "operationId": "update_bill_api_v1_bills__bill_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "bill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bill Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Bill",
        "tags": [
          "bills"
        ]
      }
    },
    "/api/v1/budgets": {
      "get": {
        "operationId": "list_budgets_api_v1_budgets_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BudgetOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Budgets",
        "tags": [
          "budgets"
        ]
      },
      "post": {
        "operationId": "create_budget_api_v1_budgets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Budget",
        "tags": [
          "budgets"
        ]
      }
    },
    "/api/v1/budgets/progress": {
      "get": {
        "description": "Gasto de cada orcamento no periodo que contem a data `on` (hoje, se omitida).",
        "operationId": "budgets_progress_api_v1_budgets_progress_get",
        "parameters": [
          {
            "in": "query",
            "name": "on",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "On"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BudgetProgressOut"
                  },
                  "title": "Response Budgets Progress Api V1 Budgets Progress Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Budgets Progress",
        "tags": [
          "budgets"
        ]
      }
    },
    "/api/v1/budgets/{budget_id}": {
      "delete": {
        "operationId": "delete_budget_api_v1_budgets__budget_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Budget Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Budget",
        "tags": [
          "budgets"
        ]
      },
      "get": {
        "operationId": "get_budget_api_v1_budgets__budget_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Budget Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Budget",
        "tags": [
          "budgets"
        ]
      },
      "patch": {
        "operationId": "update_budget_api_v1_budgets__budget_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Budget Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Budget",
        "tags": [
          "budgets"
        ]
      }
    },
...
  }
}